added alerts and restrictions to ranges and names filter

+++ src/customExceptions/InvalidRangeException.java
@@ -3,10 +3,10 @@ package customExceptions;
 public class InvalidRangeException extends Exception{
 	private static final long serialVersionUID = 1L;
 	private String name;
-	private int min;
-	private int max;
+	private double min;
+	private double max;
 	
-	public InvalidRangeException(int min, int max, String name){
+	public InvalidRangeException(double min, double max, String name){
 		this.name = name;
 		this.min= min;
 		this.max = max;
